added color for date picker material
